Extract documents table into its own component
