+++ src/components/ResidentDocumentsTable.tsx
@@ -1,0 +1,271 @@
+import React, { FunctionComponent, useMemo, useState } from 'react';
+import { EvidenceList, EvidenceTile } from './EvidenceTile';
+import { formatDate } from '../helpers/formatters';
+import { EvidenceAwaitingSubmissionTile } from './EvidenceAwaitingSubmissionTile';
+import { EvidenceRequest } from '../domain/evidence-request';
+import { DateTime } from 'luxon';
+import {
+  DocumentSubmission,
+  IDocumentSubmission,
+} from '../domain/document-submission';
+import { DocumentType } from '../domain/document-type';
+import { useRouter } from 'next/router';
+
+type EvidenceAwaitingSubmission = {
+  documentType: string;
+  dateRequested: DateTime | undefined;
+  requestedBy: string | undefined;
+  reason: string | undefined;
+  kind: 'EvidenceAwaitingSubmission';
+};
+
+type DocumentSubmissionWithKind = IDocumentSubmission & {
+  kind: 'DocumentSubmissionWithKind';
+};
+
+type DocumentTabItem = DocumentSubmissionWithKind | EvidenceAwaitingSubmission;
+
+type DocumentTab = {
+  id: string;
+  humanReadableName: string;
+  documents: DocumentTabItem[];
+  className: 'govuk-form-group--error' | undefined;
+  style: '#FFF6BB' | '#8EB6DC' | '#B2DFDB' | '#F8D1CD' | undefined;
+};
+
+export const ResidentDocumentsTable: FunctionComponent<Props> = ({
+  evidenceRequests,
+  documentSubmissions,
+  teamId,
+}) => {
+  const [selectedTab, setSelectedTab] = useState('all-documents');
+  const selectTab = (tabName: string) => {
+    if (selectedTab === tabName) {
+      return 'govuk-tabs__list-item  govuk-tabs__list-item--selected';
+    } else return 'govuk-tabs__list-item';
+  };
+  const handleTabClick = (tab: string) => {
+    setSelectedTab(tab);
+  };
+
+  const showPanel = (tabName: string) => {
+    if (selectedTab === tabName) {
+      return 'govuk-tabs__panel';
+    } else return 'govuk-tabs__panel govuk-tabs__panel--hidden';
+  };
+
+  const router = useRouter();
+  const { residentId } = router.query as {
+    residentId: string;
+  };
+
+  const toReviewDocumentSubmissions = documentSubmissions
+    .filter((ds) => ds.state == 'UPLOADED' && ds.document?.fileType)
+    .map<DocumentSubmissionWithKind>((ds) => ({
+      ...ds,
+      kind: 'DocumentSubmissionWithKind',
+    }));
+  const reviewedDocumentSubmissions = documentSubmissions
+    .filter((ds) => ds.state == 'APPROVED')
+    .map<DocumentSubmissionWithKind>((ds) => ({
+      ...ds,
+      kind: 'DocumentSubmissionWithKind',
+    }));
+  const rejectedDocumentSubmissions = documentSubmissions
+    .filter((ds) => ds.state == 'REJECTED')
+    .map<DocumentSubmissionWithKind>((ds) => ({
+      ...ds,
+      kind: 'DocumentSubmissionWithKind',
+    }));
+  const evidenceAwaitingSubmissions = useMemo(() => {
+    const documentTypesMap = new Map<string, Set<DocumentType>>();
+    evidenceRequests.forEach((er) =>
+      documentTypesMap.set(er.id, new Set(er.documentTypes))
+    );
+    documentSubmissions.forEach((ds) => {
+      const currentDocumentTypesSet = documentTypesMap.get(
+        ds.evidenceRequestId
+      );
+
+      currentDocumentTypesSet?.forEach((dt) => {
+        if (dt.id === ds.documentType.id) {
+          currentDocumentTypesSet.delete(dt);
+        }
+      });
+    });
+
+    const awaitingSubmissions: EvidenceAwaitingSubmission[] = [];
+    documentTypesMap.forEach((value, key) => {
+      value.forEach((dt) => {
+        const evidenceRequestFromKey = evidenceRequests.find(
+          (er) => er.id == key
+        );
+        awaitingSubmissions.push({
+          documentType: dt.title,
+          dateRequested: evidenceRequestFromKey?.createdAt,
+          requestedBy: evidenceRequestFromKey?.userRequestedBy,
+          reason: evidenceRequestFromKey?.reason,
+          kind: 'EvidenceAwaitingSubmission',
+        });
+      });
+    });
+    return awaitingSubmissions;
+  }, [evidenceRequests, documentSubmissions]);
+  const allDocumentSubmissions = [
+    ...toReviewDocumentSubmissions,
+    ...reviewedDocumentSubmissions,
+    ...rejectedDocumentSubmissions,
+    ...evidenceAwaitingSubmissions,
+  ];
+
+  const DocumentTabs: DocumentTab[] = [
+    {
+      id: 'all-documents',
+      humanReadableName: 'All documents',
+      documents: allDocumentSubmissions,
+      className: undefined,
+      style: undefined,
+    },
+    {
+      id: 'awaiting-submission',
+      humanReadableName: 'Awaiting submission',
+      documents: evidenceAwaitingSubmissions,
+      className: 'govuk-form-group--error',
+      style: '#FFF6BB',
+    },
+    {
+      id: 'pending-review',
+      humanReadableName: 'Pending review',
+      documents: toReviewDocumentSubmissions,
+      className: 'govuk-form-group--error',
+      style: '#8EB6DC',
+    },
+    {
+      id: 'approved',
+      humanReadableName: 'Approved',
+      documents: reviewedDocumentSubmissions,
+      className: 'govuk-form-group--error',
+      style: '#B2DFDB',
+    },
+    {
+      id: 'rejected',
+      humanReadableName: 'Rejected',
+      documents: rejectedDocumentSubmissions,
+      className: 'govuk-form-group--error',
+      style: '#F8D1CD',
+    },
+  ];
+
+  const getReason = (id: string) => {
+    const evidenceRequest = evidenceRequests.find((er) => er.id === id);
+    return evidenceRequest?.reason;
+  };
+
+  const getUserRequestedBy = (id: string) => {
+    const evidenceRequest = evidenceRequests.find((er) => er.id === id);
+    return evidenceRequest?.userRequestedBy;
+  };
+
+  return (
+    <div
+      className="js-enabled"
+      style={{
+        paddingTop: '1.5em',
+      }}
+    >
+      <div className="govuk-tabs lbh-tabs" data-module="govuk-tabs">
+        <ul className="govuk-tabs__list">
+          {DocumentTabs.map((documentTab) => {
+            return (
+              <li className={selectTab(documentTab.id)}>
+                <a
+                  className="govuk-tabs__tab"
+                  onClick={() => handleTabClick(documentTab.id)}
+                  href={'#' + documentTab.id}
+                >
+                  <h2 className="govuk-body">
+                    {documentTab.humanReadableName}
+                  </h2>
+                </a>
+              </li>
+            );
+          })}
+        </ul>
+        {DocumentTabs.map((documentTab) => {
+          return (
+            <section className={showPanel(documentTab.id)} id={documentTab.id}>
+              <article
+                className={documentTab.className}
+                style={{ borderLeftColor: documentTab.style }}
+              >
+                <EvidenceList>
+                  {documentTab.documents && documentTab.documents.length > 0 ? (
+                    documentTab.documents.map(
+                      (documentTabItem: DocumentTabItem, index) => {
+                        switch (documentTabItem.kind) {
+                          case 'DocumentSubmissionWithKind':
+                            return (
+                              <EvidenceTile
+                                teamId={teamId}
+                                residentId={residentId}
+                                key={documentTabItem.id}
+                                id={documentTabItem.id}
+                                title={
+                                  documentTabItem.staffSelectedDocumentType
+                                    ?.title ||
+                                  documentTabItem.documentType.title
+                                }
+                                createdAt={formatDate(
+                                  documentTabItem.createdAt
+                                )}
+                                fileSizeInBytes={
+                                  documentTabItem.document
+                                    ? documentTabItem.document.fileSizeInBytes
+                                    : 0
+                                }
+                                format={
+                                  documentTabItem.document
+                                    ? documentTabItem.document.extension
+                                    : 'unknown'
+                                }
+                                state={documentTabItem.state}
+                                reason={getReason(
+                                  documentTabItem.evidenceRequestId
+                                )}
+                                requestedBy={getUserRequestedBy(
+                                  documentTabItem.evidenceRequestId
+                                )}
+                                userUpdatedBy={documentTabItem.userUpdatedBy}
+                              />
+                            );
+                          case 'EvidenceAwaitingSubmission':
+                            return (
+                              <EvidenceAwaitingSubmissionTile
+                                id={index}
+                                documentType={documentTabItem.documentType}
+                                dateRequested={documentTabItem.dateRequested}
+                                requestedBy={documentTabItem.requestedBy}
+                                reason={documentTabItem.reason}
+                              />
+                            );
+                        }
+                      }
+                    )
+                  ) : (
+                    <h3>There are no documents to review</h3>
+                  )}
+                </EvidenceList>
+              </article>
+            </section>
+          );
+        })}
+      </div>
+    </div>
+  );
+};
+
+interface Props {
+  evidenceRequests: EvidenceRequest[];
+  documentSubmissions: DocumentSubmission[];
+  teamId: string;
+}
